+++ redash/settings/__init__.py
@@ -189,7 +189,7 @@ default_query_runners = [
     'redash.query_runner.db2',
     'redash.query_runner.druid',
     'redash.query_runner.kylin',
-    'redash.query_runner.drill'
+    'redash.query_runner.drill',
 ]
 
 enabled_query_runners = array_from_string(os.environ.get("REDASH_ENABLED_QUERY_RUNNERS", ",".join(default_query_runners)))
