+++ tests/query_runner/test_http.py
@@ -41,7 +41,7 @@ class TestBaseHTTPQueryRunner(TestCase):
             query_runner.get_auth
         )
 
-    @mock.patch('requests.get')
+    @mock.patch('requests.request')
     def test_get_response_success(self, mock_get):
         mock_response = mock.Mock()
         mock_response.status_code = 200
@@ -51,11 +51,11 @@ class TestBaseHTTPQueryRunner(TestCase):
         url = 'https://example.com/'
         query_runner = BaseHTTPQueryRunner({})
         response, error = query_runner.get_response(url)
-        mock_get.assert_called_once_with(url, auth=None)
+        mock_get.assert_called_once_with('get', url, auth=None)
         self.assertEqual(response.status_code, 200)
         self.assertIsNone(error)
 
-    @mock.patch('requests.get')
+    @mock.patch('requests.request')
     def test_get_response_success_custom_auth(self, mock_get):
         mock_response = mock.Mock()
         mock_response.status_code = 200
@@ -66,11 +66,11 @@ class TestBaseHTTPQueryRunner(TestCase):
         query_runner = BaseHTTPQueryRunner({})
         auth = ('username', 'password')
         response, error = query_runner.get_response(url, auth=auth)
-        mock_get.assert_called_once_with(url, auth=auth)
+        mock_get.assert_called_once_with('get', url, auth=auth)
         self.assertEqual(response.status_code, 200)
         self.assertIsNone(error)
 
-    @mock.patch('requests.get')
+    @mock.patch('requests.request')
     def test_get_response_failure(self, mock_get):
         mock_response = mock.Mock()
         mock_response.status_code = 301
@@ -80,10 +80,10 @@ class TestBaseHTTPQueryRunner(TestCase):
         url = 'https://example.com/'
         query_runner = BaseHTTPQueryRunner({})
         response, error = query_runner.get_response(url)
-        mock_get.assert_called_once_with(url, auth=None)
+        mock_get.assert_called_once_with('get', url, auth=None)
         self.assertIn(query_runner.response_error, error)
 
-    @mock.patch('requests.get')
+    @mock.patch('requests.request')
     def test_get_response_httperror_exception(self, mock_get):
         mock_response = mock.Mock()
         mock_response.status_code = 500
@@ -95,11 +95,11 @@ class TestBaseHTTPQueryRunner(TestCase):
         url = 'https://example.com/'
         query_runner = BaseHTTPQueryRunner({})
         response, error = query_runner.get_response(url)
-        mock_get.assert_called_once_with(url, auth=None)
+        mock_get.assert_called_once_with('get', url, auth=None)
         self.assertIsNotNone(error)
         self.assertIn("Failed to execute query", error)
 
-    @mock.patch('requests.get')
+    @mock.patch('requests.request')
     def test_get_response_requests_exception(self, mock_get):
         mock_response = mock.Mock()
         mock_response.status_code = 500
@@ -112,11 +112,11 @@ class TestBaseHTTPQueryRunner(TestCase):
         url = 'https://example.com/'
         query_runner = BaseHTTPQueryRunner({})
         response, error = query_runner.get_response(url)
-        mock_get.assert_called_once_with(url, auth=None)
+        mock_get.assert_called_once_with('get', url, auth=None)
         self.assertIsNotNone(error)
         self.assertEqual(exception_message, error)
 
-    @mock.patch('requests.get')
+    @mock.patch('requests.request')
     def test_get_response_generic_exception(self, mock_get):
         mock_response = mock.Mock()
         mock_response.status_code = 500
